requirements.txt
streamlit
numpy
pandas
scikit-learn
openai
google-generativeai
